requirements.txt
speechrecognition
pyttsx3
openai
pyaudio
python-dotenv
